prisma service refactor
