+++ prisma/prisma.module.ts
@@ -1,9 +1,9 @@
-import { Global, Module } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
+// import { Global, Module } from '@nestjs/common';
+// import { PrismaService } from './prisma.service';
 
-@Global()
-@Module({
-  providers: [PrismaService],
-  exports: [PrismaService],
-})
-export class PrismaModule {}
+// @Global()
+// @Module({
+//   providers: [PrismaService],
+//   exports: [PrismaService],
+// })
+// export class PrismaModule {}
