+++ prisma/prisma.service.ts
@@ -1,15 +1,15 @@
-import { INestApplication, Injectable, OnModuleInit } from '@nestjs/common';
-import { PrismaClient } from '@prisma/client';
+// import { INestApplication, Injectable, OnModuleInit } from '@nestjs/common';
+// import { PrismaClient } from '@prisma/client';
 
-@Injectable()
-export class PrismaService extends PrismaClient implements OnModuleInit {
-  async onModuleInit() {
-    await this.$connect();
-  }
+// @Injectable()
+// export class PrismaService extends PrismaClient implements OnModuleInit {
+//   async onModuleInit() {
+//     await this.$connect();
+//   }
 
-  async enableShutdownHooks(app: INestApplication) {
-    this.$on('beforeExit', async () => {
-      await app.close();
-    });
-  }
-}
+//   async enableShutdownHooks(app: INestApplication) {
+//     this.$on('beforeExit', async () => {
+//       await app.close();
+//     });
+//   }
+// }
